Tweaked compile-definition flags and build settings
